Missing test edit class
