+++ step_definitions/management_classes/createClass_steps.js
@@ -1,6 +1,7 @@
 
 const navigationLeftPanel = require('../../page-objects/pages/NavigationLeftPage');
-const manageClasses = require('../../page-objects/pages/managementPages/ManageClassesPage');
+const manageClass = require('../../page-objects/pages/managementPages/ManageClassPage');
+const manageClassesPage = require('../../page-objects/pages/managementPages/ManageClassesPage');
 
 When(/^.* goes to management$/, async function () {
 	await navigationLeftPanel.clickNavItemManagement();
@@ -11,17 +12,17 @@ Given(/^.* goes to class management$/, async function () {
 });
 
 When(/^.* creates class with custom name '([^']*)'$/, async function (customClassName) {
-	await manageClasses.createNewClass({customClassName: customClassName});
+	await manageClass.createNewClass({customClassName: customClassName});
 });
 
 When(/^.* creates a class with grade '([^']*)' and name '([^']*)'$/, async function (className, classGrade) {
-	await manageClasses.createNewClass({className: className, classGrade: classGrade});
+	await manageClass.createNewClass({className: className, classGrade: classGrade});
 });
 
 When(/^.* opens classes tab with name '([^']*)'$/, async function (classesTabName) {
-	await manageClasses.clickOnClassesTab(classesTabName);
+	await manageClassesPage.clickOnClassesTab(classesTabName);
 });
 
 Then(/^.* class with name '([^']*)' and '([^']*)' members is visible$/, async function (className, membersCount) {
-	await manageClasses.isNewEmptyClassCreated(className, membersCount);
+	await manageClass.isNewEmptyClassCreated(className, membersCount);
 });
